+++ tests/factories/main.php
@@ -1,6 +1,6 @@
 <?php
 
-use League\FactoryMuffin\Facade\FactoryMuffin;
+use League\FactoryMuffin\Facade as FactoryMuffin;
 
 FactoryMuffin::define('IdTestModelGetKeyStub', array());
 FactoryMuffin::define('IdTestModelPkStub', array());
