+++ tests/factories/savedelete.php
@@ -1,6 +1,6 @@
 <?php
 
-use League\FactoryMuffin\Facade\FactoryMuffin;
+use League\FactoryMuffin\Facade as FactoryMuffin;
 
 FactoryMuffin::define('ModelThatWillSaveStub', array());
 FactoryMuffin::define('ModelThatFailsToSaveStub', array());
