+++ tests/stubs/example.php
@@ -1,5 +1,5 @@
 <?php
 
-use League\FactoryMuffin\Facade\FactoryMuffin;
+use League\FactoryMuffin\Facade as FactoryMuffin;
 
 FactoryMuffin::define('ExampleDefinedModelStub', array());
